split BSScript into files
