+++ CommonLibF4/include/RE/Bethesda/BSScript/Object.h
@@ -1,0 +1,64 @@
+#pragma once
+
+#include "RE/Bethesda/BSContainer.h"
+#include "RE/Bethesda/BSFixedString.h"
+#include "RE/Bethesda/BSLock.h"
+#include "RE/Bethesda/BSTArray.h"
+#include "RE/Bethesda/BSTEvent.h"
+#include "RE/Bethesda/BSTHashMap.h"
+#include "RE/Bethesda/BSTObjectArena.h"
+#include "RE/Bethesda/BSTSmartPointer.h"
+#include "RE/Bethesda/BSTTuple.h"
+#include "RE/Bethesda/MemoryManager.h"
+#include "RE/Bethesda/BSScript/ObjectTypeInfo.h"
+#include "RE/Bethesda/BSScript/Variable.h"
+namespace RE
+{
+	namespace BSScript
+	{
+		class ObjectTypeInfo;
+		class Variable;
+
+		class Object
+		{
+		public:
+			~Object();
+
+			[[nodiscard]] bool Constructed() const noexcept { return constructed; }
+
+			[[nodiscard]] std::uint32_t DecRef() const
+			{
+				using func_t = decltype(&Object::DecRef);
+				REL::Relocation<func_t> func{ REL::ID(541793) };
+				return func(this);
+			}
+
+			[[nodiscard]] std::size_t GetHandle() const
+			{
+				using func_t = decltype(&Object::GetHandle);
+				REL::Relocation<func_t> func{ REL::ID(1452752) };
+				return func(this);
+			}
+
+			void IncRef() const
+			{
+				using func_t = decltype(&Object::IncRef);
+				REL::Relocation<func_t> func{ REL::ID(461710) };
+				return func(this);
+			}
+
+			// members
+			std::uint32_t constructed: 1;            // 00:00
+			std::uint32_t initialized: 1;            // 00:01
+			std::uint32_t valid: 1;                  // 00:02
+			std::uint32_t remainingPropsToInit: 29;  // 00:03
+			BSTSmartPointer<ObjectTypeInfo> type;    // 08
+			BSFixedString currentState;              // 10
+			void* lockStructure;                     // 18
+			std::size_t handle;                      // 20
+			std::uint32_t refCountAndHandleLock;     // 28
+			Variable variables[0];                   // 30
+		};
+		static_assert(sizeof(Object) == 0x30);
+	}
+}
