+++ CommonLibF4/include/RE/Bethesda/BSScript/PropertyGroupInfo.h
@@ -1,0 +1,30 @@
+#pragma once
+
+#include "RE/Bethesda/BSContainer.h"
+#include "RE/Bethesda/BSFixedString.h"
+#include "RE/Bethesda/BSLock.h"
+#include "RE/Bethesda/BSTArray.h"
+#include "RE/Bethesda/BSTEvent.h"
+#include "RE/Bethesda/BSTHashMap.h"
+#include "RE/Bethesda/BSTObjectArena.h"
+#include "RE/Bethesda/BSTSmartPointer.h"
+#include "RE/Bethesda/BSTTuple.h"
+#include "RE/Bethesda/MemoryManager.h"
+namespace RE
+{
+	namespace BSScript
+	{
+
+		class PropertyGroupInfo :
+			public BSIntrusiveRefCounted  // 00
+		{
+		public:
+			// members
+			BSFixedString groupName;                // 08
+			BSFixedString docString;                // 10
+			std::uint32_t userFlags;                // 18
+			BSTArray<BSFixedString> propertyNames;  // 20
+		};
+		static_assert(sizeof(PropertyGroupInfo) == 0x38);
+	}
+}
