+++ CommonLibF4/include/RE/Fallout.h
@@ -42,6 +42,37 @@
 #include "RE/Bethesda/BSResource.h"
 #include "RE/Bethesda/BSScaleformManager.h"
 #include "RE/Bethesda/BSScript.h"
+#include "RE/Bethesda/BSScript/Array.h"
+#include "RE/Bethesda/BSScript/ArrayWrapper.h"
+#include "RE/Bethesda/BSScript/CompiledScriptLoader.h"
+#include "RE/Bethesda/BSScript/ErrorLogger.h"
+#include "RE/Bethesda/BSScript/ICachedErrorMessage.h"
+#include "RE/Bethesda/BSScript/IClientVM.h"
+#include "RE/Bethesda/BSScript/IComplexType.h"
+#include "RE/Bethesda/BSScript/IFunction.h"
+#include "RE/Bethesda/BSScript/ILoader.h"
+#include "RE/Bethesda/BSScript/IMemoryPagePolicy.h"
+#include "RE/Bethesda/BSScript/IObjectHandlePolicy.h"
+#include "RE/Bethesda/BSScript/IProfilePolicy.h"
+#include "RE/Bethesda/BSScript/ISavePatcherInterface.h"
+#include "RE/Bethesda/BSScript/IStackCallbackFunctor.h"
+#include "RE/Bethesda/BSScript/IStackCallbackSaveInterface.h"
+#include "RE/Bethesda/BSScript/IVirtualMachine.h"
+#include "RE/Bethesda/BSScript/Internal/ReadableStringTable.h"
+#include "RE/Bethesda/BSScript/Internal/VDescTable.h"
+#include "RE/Bethesda/BSScript/MergedBoundScript.h"
+#include "RE/Bethesda/BSScript/NF_util/NativeFunctionBase.h"
+#include "RE/Bethesda/BSScript/Object.h"
+#include "RE/Bethesda/BSScript/ObjectBindPolicy.h"
+#include "RE/Bethesda/BSScript/ObjectTypeInfo.h"
+#include "RE/Bethesda/BSScript/PropertyGroupInfo.h"
+#include "RE/Bethesda/BSScript/SimpleAllocMemoryPagePolicy.h"
+#include "RE/Bethesda/BSScript/Stack.h"
+#include "RE/Bethesda/BSScript/StackFrame.h"
+#include "RE/Bethesda/BSScript/Struct.h"
+#include "RE/Bethesda/BSScript/StructTypeInfo.h"
+#include "RE/Bethesda/BSScript/TypeInfo.h"
+#include "RE/Bethesda/BSScript/Variable.h"
 #include "RE/Bethesda/BSScriptUtil.h"
 #include "RE/Bethesda/BSSemaphore.h"
 #include "RE/Bethesda/BSShader.h"
